backend/requirements.txt
asgiref==3.7.2
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.2.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chromadb
Django==4.2.3
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-reversion==5.0.12
django-tabular-permissions==2.9.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.5.0
langchain
langchain-core
langchain-mistralai
langchain-experimental
langchain-openai
langgraph==0.0.51
pysqlite3-binary
langsmith==0.1.92
numpy==1.26.1
openai==1.35.12
PyJWT==2.8.0
pypdf
PyPDF2
pyOpenSSL==23.3.0
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.25.0
psycopg2
psycopg[binary]==3.1.12
PyMuPDF
tiktoken
langfuse
chainlit
